requirements.txt
matplotlib==3.9.4
pandas==2.2.3
pyserial==3.5
